add export btn

+++ database/migrations/2022_01_27_214146_add_quantite_to_commandes_table.php
@@ -14,7 +14,7 @@ class AddQuantiteToCommandesTable extends Migration
     public function up()
     {
         Schema::table('commandes', function (Blueprint $table) {
-            $table->integer('quantite')->nullable();
+            $table->float('quantite')->nullable();
         });
     }
